Added a sound driver optimisation
Reading through
https://github.com/nukeykt/Nuked-OPN2/blob/master/ym3438.c (a
translation of a YM3438 die shot to C) gives some insight on why this
works:

If I'm understanding it right, while the YM malfunctions if both an
address port and a data port are written to without a YM cycle occuring
in between (both writes get stored in the same place), a YM cycle occurs
every 2.8 Z80 cycles. Since the 'ld a,c' instruction takes 4 Z80 cycles,
and the other 'ld' instructions take 13 cycles, there's more than enough
time for the YM to catch the writes.